+++ src/main/java/nextstep/security/authentication/AuthenticationSuccessHandler.java
@@ -1,11 +1,11 @@
-package nextstep.security.oauth2;
+package nextstep.security.authentication;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.authentication.Authentication;
 
 import java.io.IOException;
 
+@FunctionalInterface
 public interface AuthenticationSuccessHandler {
 
     void onSuccess(final HttpServletRequest request, final HttpServletResponse response, Authentication authentication) throws IOException;
